requirements.txt
sentence-transformers
faiss-cpu
langgraph
langchain-groq